Prevent getting banned for "ping spamming"
The Group owner no longer can add others to the Group.  The owner
creates the Group and then every person who wants in has to `addme`
themselves.  The `addme` command will track the Reddit permalink where
they asked to be added as proof they added themselves.

Because of this, Groups are allowed to have no Members and removing all
Members no longer auto-deletes the Group.